requirements.txt
numpy==1.26.2
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
termcolor==3.1.0
omegaconf==2.3.0
loguru==0.7.2
jaxtyping==0.2.25
imageio==2.33.1
imageio-ffmpeg==0.4.9
matplotlib==3.9.4
pandas==2.2.3
h5py==3.10.0
open3d==0.18.0
trimesh==4.0.5
PyMCubes==0.1.6
diso==0.1.4
lpips==0.1.4
einops==0.8.0
rembg==2.0.66
ninja==1.11.1.3
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast/
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d
transformers==4.36.0

